+++ drivers/net/ethernet/intel/idpf/idpf_txrx.h
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
-/* Copyright (C) 2023 Intel Corporation */
+/* Copyright (C) 2024 Intel Corporation */
 
 #ifndef _IDPF_TXRX_H_
 #define _IDPF_TXRX_H_
