+++ drivers/net/ethernet/intel/idpf/idpf_virtchnl.c
@@ -6,63 +6,6 @@
 #include "idpf_virtchnl.h"
 
 /**
- * idpf_vid_to_vport - Translate vport id to vport pointer
- * @adapter: private data struct
- * @v_id: vport id to translate
- *
- * Returns vport matching v_id, NULL if not found.
- */
-struct idpf_vport *idpf_vid_to_vport(struct idpf_adapter *adapter, u32 v_id)
-{
-	u16 num_max_vports = idpf_get_max_vports(adapter);
-	int i;
-
-	for (i = 0; i < num_max_vports; i++)
-		if (adapter->vport_ids[i] == v_id)
-			return adapter->vports[i];
-
-	return NULL;
-}
-
-/**
- * idpf_handle_event_link - Handle link event message
- * @adapter: private data struct
- * @v2e: virtchnl event message
- */
-static void idpf_handle_event_link(struct idpf_adapter *adapter,
-				   const struct virtchnl2_event *v2e)
-{
-	struct idpf_netdev_priv *np;
-	struct idpf_vport *vport;
-
-	vport = idpf_vid_to_vport(adapter, le32_to_cpu(v2e->vport_id));
-	if (!vport) {
-		dev_err_ratelimited(&adapter->pdev->dev, "Failed to find vport_id %d for link event\n",
-				    v2e->vport_id);
-		return;
-	}
-	np = netdev_priv(vport->netdev);
-
-	vport->link_speed_mbps = le32_to_cpu(v2e->link_speed);
-
-	if (vport->link_up == v2e->link_status)
-		return;
-
-	vport->link_up = v2e->link_status;
-
-	if (np->state != __IDPF_VPORT_UP)
-		return;
-
-	if (vport->link_up) {
-		netif_tx_start_all_queues(vport->netdev);
-		netif_carrier_on(vport->netdev);
-	} else {
-		netif_tx_stop_all_queues(vport->netdev);
-		netif_carrier_off(vport->netdev);
-	}
-}
-
-/**
  * idpf_recv_event_msg - Receive virtchnl event message
  * @adapter: Driver specific private structure
  * @ctlq_msg: message to copy from
@@ -943,188 +886,6 @@ int idpf_get_reg_intr_vecs(struct idpf_vport *vport,
 	}
 
 	return num_regs;
-}
-
-/**
- * idpf_vport_get_q_reg - Get the queue registers for the vport
- * @reg_vals: register values needing to be set
- * @num_regs: amount we expect to fill
- * @q_type: queue model
- * @chunks: queue regs received over mailbox
- *
- * This function parses the queue register offsets from the queue register
- * chunk information, with a specific queue type and stores it into the array
- * passed as an argument. It returns the actual number of queue registers that
- * are filled.
- */
-static int idpf_vport_get_q_reg(u32 *reg_vals, int num_regs, u32 q_type,
-				struct virtchnl2_queue_reg_chunks *chunks)
-{
-	u16 num_chunks = le16_to_cpu(chunks->num_chunks);
-	int reg_filled = 0, i;
-	u32 reg_val;
-
-	while (num_chunks--) {
-		struct virtchnl2_queue_reg_chunk *chunk;
-		u16 num_q;
-
-		chunk = &chunks->chunks[num_chunks];
-		if (le32_to_cpu(chunk->type) != q_type)
-			continue;
-
-		num_q = le32_to_cpu(chunk->num_queues);
-		reg_val = le64_to_cpu(chunk->qtail_reg_start);
-		for (i = 0; i < num_q && reg_filled < num_regs ; i++) {
-			reg_vals[reg_filled++] = reg_val;
-			reg_val += le32_to_cpu(chunk->qtail_reg_spacing);
-		}
-	}
-
-	return reg_filled;
-}
-
-/**
- * __idpf_queue_reg_init - initialize queue registers
- * @vport: virtual port structure
- * @reg_vals: registers we are initializing
- * @num_regs: how many registers there are in total
- * @q_type: queue model
- *
- * Return number of queues that are initialized
- */
-static int __idpf_queue_reg_init(struct idpf_vport *vport, u32 *reg_vals,
-				 int num_regs, u32 q_type)
-{
-	struct idpf_adapter *adapter = vport->adapter;
-	struct idpf_queue *q;
-	int i, j, k = 0;
-
-	switch (q_type) {
-	case VIRTCHNL2_QUEUE_TYPE_TX:
-		for (i = 0; i < vport->num_txq_grp; i++) {
-			struct idpf_txq_group *tx_qgrp = &vport->txq_grps[i];
-
-			for (j = 0; j < tx_qgrp->num_txq && k < num_regs; j++, k++)
-				tx_qgrp->txqs[j]->tail =
-					idpf_get_reg_addr(adapter, reg_vals[k]);
-		}
-		break;
-	case VIRTCHNL2_QUEUE_TYPE_RX:
-		for (i = 0; i < vport->num_rxq_grp; i++) {
-			struct idpf_rxq_group *rx_qgrp = &vport->rxq_grps[i];
-			u16 num_rxq = rx_qgrp->singleq.num_rxq;
-
-			for (j = 0; j < num_rxq && k < num_regs; j++, k++) {
-				q = rx_qgrp->singleq.rxqs[j];
-				q->tail = idpf_get_reg_addr(adapter,
-							    reg_vals[k]);
-			}
-		}
-		break;
-	case VIRTCHNL2_QUEUE_TYPE_RX_BUFFER:
-		for (i = 0; i < vport->num_rxq_grp; i++) {
-			struct idpf_rxq_group *rx_qgrp = &vport->rxq_grps[i];
-			u8 num_bufqs = vport->num_bufqs_per_qgrp;
-
-			for (j = 0; j < num_bufqs && k < num_regs; j++, k++) {
-				q = &rx_qgrp->splitq.bufq_sets[j].bufq;
-				q->tail = idpf_get_reg_addr(adapter,
-							    reg_vals[k]);
-			}
-		}
-		break;
-	default:
-		break;
-	}
-
-	return k;
-}
-
-/**
- * idpf_queue_reg_init - initialize queue registers
- * @vport: virtual port structure
- *
- * Return 0 on success, negative on failure
- */
-int idpf_queue_reg_init(struct idpf_vport *vport)
-{
-	struct virtchnl2_create_vport *vport_params;
-	struct virtchnl2_queue_reg_chunks *chunks;
-	struct idpf_vport_config *vport_config;
-	u16 vport_idx = vport->idx;
-	int num_regs, ret = 0;
-	u32 *reg_vals;
-
-	/* We may never deal with more than 256 same type of queues */
-	reg_vals = kzalloc(sizeof(void *) * IDPF_LARGE_MAX_Q, GFP_KERNEL);
-	if (!reg_vals)
-		return -ENOMEM;
-
-	vport_config = vport->adapter->vport_config[vport_idx];
-	if (vport_config->req_qs_chunks) {
-		struct virtchnl2_add_queues *vc_aq =
-		  (struct virtchnl2_add_queues *)vport_config->req_qs_chunks;
-		chunks = &vc_aq->chunks;
-	} else {
-		vport_params = vport->adapter->vport_params_recvd[vport_idx];
-		chunks = &vport_params->chunks;
-	}
-
-	/* Initialize Tx queue tail register address */
-	num_regs = idpf_vport_get_q_reg(reg_vals, IDPF_LARGE_MAX_Q,
-					VIRTCHNL2_QUEUE_TYPE_TX,
-					chunks);
-	if (num_regs < vport->num_txq) {
-		ret = -EINVAL;
-		goto free_reg_vals;
-	}
-
-	num_regs = __idpf_queue_reg_init(vport, reg_vals, num_regs,
-					 VIRTCHNL2_QUEUE_TYPE_TX);
-	if (num_regs < vport->num_txq) {
-		ret = -EINVAL;
-		goto free_reg_vals;
-	}
-
-	/* Initialize Rx/buffer queue tail register address based on Rx queue
-	 * model
-	 */
-	if (idpf_is_queue_model_split(vport->rxq_model)) {
-		num_regs = idpf_vport_get_q_reg(reg_vals, IDPF_LARGE_MAX_Q,
-						VIRTCHNL2_QUEUE_TYPE_RX_BUFFER,
-						chunks);
-		if (num_regs < vport->num_bufq) {
-			ret = -EINVAL;
-			goto free_reg_vals;
-		}
-
-		num_regs = __idpf_queue_reg_init(vport, reg_vals, num_regs,
-						 VIRTCHNL2_QUEUE_TYPE_RX_BUFFER);
-		if (num_regs < vport->num_bufq) {
-			ret = -EINVAL;
-			goto free_reg_vals;
-		}
-	} else {
-		num_regs = idpf_vport_get_q_reg(reg_vals, IDPF_LARGE_MAX_Q,
-						VIRTCHNL2_QUEUE_TYPE_RX,
-						chunks);
-		if (num_regs < vport->num_rxq) {
-			ret = -EINVAL;
-			goto free_reg_vals;
-		}
-
-		num_regs = __idpf_queue_reg_init(vport, reg_vals, num_regs,
-						 VIRTCHNL2_QUEUE_TYPE_RX);
-		if (num_regs < vport->num_rxq) {
-			ret = -EINVAL;
-			goto free_reg_vals;
-		}
-	}
-
-free_reg_vals:
-	kfree(reg_vals);
-
-	return ret;
 }
 
 /**
@@ -2957,92 +2718,6 @@ void idpf_vc_core_deinit(struct idpf_adapter *adapter)
 	adapter->vports = NULL;
 
 	clear_bit(IDPF_VC_CORE_INIT, adapter->flags);
-}
-
-/**
- * idpf_vport_alloc_vec_indexes - Get relative vector indexes
- * @vport: virtual port data struct
- *
- * This function requests the vector information required for the vport and
- * stores the vector indexes received from the 'global vector distribution'
- * in the vport's queue vectors array.
- *
- * Return 0 on success, error on failure
- */
-int idpf_vport_alloc_vec_indexes(struct idpf_vport *vport)
-{
-	struct idpf_vector_info vec_info;
-	int num_alloc_vecs;
-
-	vec_info.num_curr_vecs = vport->num_q_vectors;
-	vec_info.num_req_vecs = max(vport->num_txq, vport->num_rxq);
-	vec_info.default_vport = vport->default_vport;
-	vec_info.index = vport->idx;
-
-	num_alloc_vecs = idpf_req_rel_vector_indexes(vport->adapter,
-						     vport->q_vector_idxs,
-						     &vec_info);
-	if (num_alloc_vecs <= 0) {
-		dev_err(&vport->adapter->pdev->dev, "Vector distribution failed: %d\n",
-			num_alloc_vecs);
-		return -EINVAL;
-	}
-
-	vport->num_q_vectors = num_alloc_vecs;
-
-	return 0;
-}
-
-/**
- * idpf_vport_init - Initialize virtual port
- * @vport: virtual port to be initialized
- * @max_q: vport max queue info
- *
- * Will initialize vport with the info received through MB earlier
- */
-void idpf_vport_init(struct idpf_vport *vport, struct idpf_vport_max_q *max_q)
-{
-	struct idpf_adapter *adapter = vport->adapter;
-	struct virtchnl2_create_vport *vport_msg;
-	struct idpf_vport_config *vport_config;
-	u16 tx_itr[] = {2, 8, 64, 128, 256};
-	u16 rx_itr[] = {2, 8, 32, 96, 128};
-	struct idpf_rss_data *rss_data;
-	u16 idx = vport->idx;
-
-	vport_config = adapter->vport_config[idx];
-	rss_data = &vport_config->user_config.rss_data;
-	vport_msg = adapter->vport_params_recvd[idx];
-
-	vport_config->max_q.max_txq = max_q->max_txq;
-	vport_config->max_q.max_rxq = max_q->max_rxq;
-	vport_config->max_q.max_complq = max_q->max_complq;
-	vport_config->max_q.max_bufq = max_q->max_bufq;
-
-	vport->txq_model = le16_to_cpu(vport_msg->txq_model);
-	vport->rxq_model = le16_to_cpu(vport_msg->rxq_model);
-	vport->vport_type = le16_to_cpu(vport_msg->vport_type);
-	vport->vport_id = le32_to_cpu(vport_msg->vport_id);
-
-	rss_data->rss_key_size = min_t(u16, NETDEV_RSS_KEY_LEN,
-				       le16_to_cpu(vport_msg->rss_key_size));
-	rss_data->rss_lut_size = le16_to_cpu(vport_msg->rss_lut_size);
-
-	ether_addr_copy(vport->default_mac_addr, vport_msg->default_mac_addr);
-	vport->max_mtu = le16_to_cpu(vport_msg->max_mtu) - IDPF_PACKET_HDR_PAD;
-
-	/* Initialize Tx and Rx profiles for Dynamic Interrupt Moderation */
-	memcpy(vport->rx_itr_profile, rx_itr, IDPF_DIM_PROFILE_SLOTS);
-	memcpy(vport->tx_itr_profile, tx_itr, IDPF_DIM_PROFILE_SLOTS);
-
-	idpf_vport_set_hsplit(vport, ETHTOOL_TCP_DATA_SPLIT_ENABLED);
-
-	idpf_vport_init_num_qs(vport, vport_msg);
-	idpf_vport_calc_num_q_desc(vport);
-	idpf_vport_calc_num_q_groups(vport);
-	idpf_vport_alloc_vec_indexes(vport);
-
-	vport->crc_enable = adapter->crc_enable;
 }
 
 /**
